Fix warnings on Linux

+++ include/RavEngine/Texture.hpp
@@ -36,7 +36,7 @@ protected:
 	bgfx::TextureHandle texture = BGFX_INVALID_HANDLE;
 	Texture(){}
 	
-	void CreateTexture(int width, int height, bool hasMipMaps, int numlayers, const uint8_t *data, int flags = BGFX_TEXTURE_SRGB | BGFX_SAMPLER_POINT);
+	void CreateTexture(int width, int height, bool hasMipMaps, int numlayers, const uint8_t *data, int flags = static_cast<int>(BGFX_TEXTURE_SRGB | BGFX_SAMPLER_POINT));
 };
 
 class RuntimeTexture : public Texture{
@@ -52,7 +52,7 @@ public:
 	 @param data pointer to the image data. Must be a 4-channel image.
 	 @param flags optional creation flags
 	 */
-	RuntimeTexture(int width, int height, bool hasMipMaps, int numlayers, const uint8_t *data, int flags = BGFX_TEXTURE_SRGB | BGFX_SAMPLER_POINT) : Texture(){
+	RuntimeTexture(int width, int height, bool hasMipMaps, int numlayers, const uint8_t *data, int flags = static_cast<int>(BGFX_TEXTURE_SRGB | BGFX_SAMPLER_POINT)) : Texture(){
 		CreateTexture(width, height, hasMipMaps, numlayers, data,flags);
 	}
 };
